+++ implementation/sequential/ShortestPrimePathSequential.cpp
@@ -7,7 +7,7 @@
 #include <cmath>
 
 // Helper function to check if a number is prime.
-bool ShortestPrimePathSequential::isPrime(int n) {
+bool ShortestPrimePathSequential::isPrime(size_t n) {
     if(n <= 1) return false;
     if(n <= 3) return true;
     if(n % 2 == 0 || n % 3 == 0) return false;
